# contracts/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Waypoint Core API",
    "version": "1.0.0",
    "summary": "Frozen v1 contracts for attributable capture, authoritative reads, standard MCP, and defensible export",
    "description": "REST resources and standard MCP delegate to one authoritative application service. Core records completed outcomes and lifecycle mutations; it never executes assessment commands. Actor and engagement scope are credential-derived."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "security": [
    {
      "actorBearer": []
    }
  ],
  "paths": {
    "/captures": {
      "post": {
        "operationId": "createCapture",
        "tags": [
          "Capture"
        ],
        "summary": "Durably ingest one completed raw-first command capture",
        "description": "The envelope and both raw streams form one idempotent request. The server acknowledges only after immutable action metadata, distinct stdout/stderr evidence references, and the capture.accepted audit event commit. A replay does not create another event.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersion"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "envelope",
                  "stdout",
                  "stderr"
                ],
                "properties": {
                  "envelope": {
                    "$ref": "#/components/schemas/CaptureEnvelope"
                  },
                  "stdout": {
                    "type": "string",
                    "format": "binary",
                    "contentMediaType": "application/octet-stream"
                  },
                  "stderr": {
                    "type": "string",
                    "format": "binary",
                    "contentMediaType": "application/octet-stream"
                  }
                }
              },
              "encoding": {
                "envelope": {
                  "contentType": "application/json"
                },
                "stdout": {
                  "contentType": "application/octet-stream"
                },
                "stderr": {
                  "contentType": "application/octet-stream"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New capture durably created",
            "headers": {
              "Waypoint-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureAck"
                }
              }
            }
          },
          "200": {
            "description": "Byte-equivalent validated capture replayed; original durable acknowledgement returned",
            "headers": {
              "Waypoint-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureAck"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "426": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/actions": {
      "get": {
        "operationId": "listActions",
        "tags": [
          "Actions"
        ],
        "summary": "Read immutable actions in bounded keyset order",
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersion"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/TechniqueFilter"
          },
          {
            "$ref": "#/components/parameters/TargetFilter"
          },
          {
            "$ref": "#/components/parameters/HostFilter"
          },
          {
            "$ref": "#/components/parameters/ActorFilter"
          },
          {
            "$ref": "#/components/parameters/ResultFilter"
          },
          {
            "$ref": "#/components/parameters/InitiatedByFilter"
          },
          {
            "$ref": "#/components/parameters/ProvenanceFilter"
          },
          {
            "$ref": "#/components/parameters/ParseStatusFilter"
          },
          {
            "$ref": "#/components/parameters/StartedAfterFilter"
          },
          {
            "$ref": "#/components/parameters/StartedBeforeFilter"
          },
          {
            "$ref": "#/components/parameters/EndedAfterFilter"
          },
          {
            "$ref": "#/components/parameters/EndedBeforeFilter"
          },
          {
            "$ref": "#/components/parameters/AfterCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful authoritative response",
            "headers": {
              "Waypoint-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "412": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "426": {
            "$ref": "#/components/responses/Problem"
          },
          "428": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          },
          "507": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "description": "Every item retains the complete accepted capture envelope and event-time actor attribution. Filters may only narrow the bounded authoritative projection."
      }
    },
...
    "/audit-events": {
      "get": {
        "operationId": "listAuditEvents",
        "tags": [
          "Audit"
        ],
        "summary": "Read persisted audit events in ascending keyset order",
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersion"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/AfterCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "A bounded page. nextCursor is exclusive on the next request.",
            "headers": {
              "Waypoint-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "426": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/events": {
      "get": {
        "operationId": "streamAuditEvents",
        "tags": [
          "Audit"
        ],
        "summary": "Resume the authorized audit stream using persisted event IDs",
        "description": "SSE id equals AuditEvent.id, SSE event equals AuditEvent.type, and SSE data is the complete AuditEvent JSON. after and Last-Event-ID are exclusive and must agree when both are sent. An absent cursor starts after the connection high-water mark. Heartbeats are comments and do not advance the cursor. Slow readers are disconnected and resume by cursor.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersion"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/AfterCursor"
          },
          {
            "$ref": "#/components/parameters/LastEventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized resumable server-sent event stream",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-cache"
                }
              },
              "Waypoint-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "x-waypoint-sse-data-schema": {
                  "$ref": "#/components/schemas/AuditEvent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "410": {
            "$ref": "#/components/responses/Problem"
          },
          "426": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/actions/needs-plugin": {
      "get": {
        "operationId": "listNeedsPluginActions",
        "tags": [
          "Actions"
        ],
        "summary": "Read needs-plugin actions in bounded keyset order",
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersion"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/TechniqueFilter"
          },
          {
            "$ref": "#/components/parameters/TargetFilter"
          },
          {
            "$ref": "#/components/parameters/HostFilter"
          },
          {
            "$ref": "#/components/parameters/ActorFilter"
          },
          {
            "$ref": "#/components/parameters/ResultFilter"
          },
          {
            "$ref": "#/components/parameters/InitiatedByFilter"
          },
          {
            "$ref": "#/components/parameters/ProvenanceFilter"
          },
          {
            "$ref": "#/components/parameters/ParseStatusFilter"
          },
          {
            "$ref": "#/components/parameters/StartedAfterFilter"
          },
          {
            "$ref": "#/components/parameters/StartedBeforeFilter"
          },
          {
            "$ref": "#/components/parameters/EndedAfterFilter"
          },
          {
            "$ref": "#/components/parameters/EndedBeforeFilter"
          },
          {
            "$ref": "#/components/parameters/AfterCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful authoritative response",
            "headers": {
              "Waypoint-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "412": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "426": {
            "$ref": "#/components/responses/Problem"
          },
          "428": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          },
          "507": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "description": "Every item retains the complete accepted capture envelope and event-time actor attribution. Filters may only narrow the bounded authoritative projection."
      }
    }
  },
  "components": {
    "securitySchemes": {
      "actorBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque actor-scoped token",
        "description": "A revocable per-actor credential. It is never accepted in a URL or request body."
      }
    },
    "parameters": {
      "ContractVersion": {
        "name": "Waypoint-Contract-Version",
        "in": "header",
        "required": true,
        "description": "Exact negotiated contract version for this request.",
        "schema": {
          "type": "string",
          "const": "1.0.0"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Must exactly equal envelope.captureId. Scope is the credential-derived engagement and actor plus sourceAgent.id and captureId.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional caller correlation value. Invalid values are replaced, never trusted for authorization.",
        "schema": {
          "$ref": "schemas/common.schema.json#/$defs/RequestId"
        }
      },
      "AfterCursor": {
        "name": "after",
        "in": "query",
        "required": false,
        "description": "Exclusive persisted audit event cursor. Omit for initial REST page or a live-only stream.",
        "schema": {
          "$ref": "#/components/schemas/Common/$defs/Cursor"
        }
      },
      "LastEventId": {
        "name": "Last-Event-ID",
        "in": "header",
        "required": false,
        "description": "Exclusive SSE resume cursor. If after is also present, values must be identical.",
        "schema": {
          "$ref": "#/components/schemas/Common/$defs/Cursor"
        }
      },
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Bounded page size.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 100
        }
      },
...
      "TechniqueFilter": {
        "name": "technique",
        "in": "query",
        "required": false,
        "description": "Exact technique / parser identifier filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "TargetFilter": {
        "name": "target",
        "in": "query",
        "required": false,
        "description": "Exact target kind or value filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "HostFilter": {
        "name": "host",
        "in": "query",
        "required": false,
        "description": "Exact executing host IP filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "ActorFilter": {
        "name": "actor",
        "in": "query",
        "required": false,
        "description": "Exact actor handle or ID filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "ResultFilter": {
        "name": "result",
        "in": "query",
        "required": false,
        "description": "Execution result filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "InitiatedByFilter": {
        "name": "initiatedBy",
        "in": "query",
        "required": false,
        "description": "Exact initiation provenance filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "ProvenanceFilter": {
        "name": "provenance",
        "in": "query",
        "required": false,
        "description": "Exact provenance bucket filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "ParseStatusFilter": {
        "name": "parseStatus",
        "in": "query",
        "required": false,
        "description": "Exact parsing status filter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        }
      },
      "StartedAfterFilter": {
        "name": "startedAfter",
        "in": "query",
        "required": false,
        "description": "Inclusive lower bound for startedAt.",
        "schema": {
          "$ref": "schemas/common.schema.json#/$defs/Timestamp"
        }
      },
      "StartedBeforeFilter": {
        "name": "startedBefore",
        "in": "query",
        "required": false,
        "description": "Inclusive upper bound for startedAt.",
        "schema": {
          "$ref": "schemas/common.schema.json#/$defs/Timestamp"
        }
      },
      "EndedAfterFilter": {
        "name": "endedAfter",
        "in": "query",
        "required": false,
        "description": "Inclusive lower bound for endedAt.",
        "schema": {
          "$ref": "schemas/common.schema.json#/$defs/Timestamp"
        }
      },
      "EndedBeforeFilter": {
        "name": "endedBefore",
        "in": "query",
        "required": false,
        "description": "Inclusive upper bound for endedAt.",
        "schema": {
          "$ref": "schemas/common.schema.json#/$defs/Timestamp"
        }
      }
    },
    "headers": {
      "ContractVersion": {
        "description": "Contract version used for the response.",
        "schema": {
          "type": "string",
          "const": "1.0.0"
        }
      },
      "RequestId": {
        "description": "Server-accepted or generated request correlation ID.",
        "schema": {
          "$ref": "schemas/common.schema.json#/$defs/RequestId"
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "A stable machine-readable problem detail. The HTTP status and body status agree.",
        "headers": {
          "Waypoint-Contract-Version": {
            "$ref": "#/components/headers/ContractVersion"
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Retry-After": {
            "description": "Present for rate or temporary capacity failures when known.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CaptureEnvelope": {
        "$ref": "schemas/capture-envelope.schema.json"
      },
      "CaptureAck": {
        "$ref": "schemas/capture-ack.schema.json"
      },
...
      "ActionPage": {
        "$ref": "schemas/action-page.schema.json"
      },
...
      "AuditEvent": {
        "$ref": "schemas/audit-event.schema.json"
      },
      "AuditPage": {
        "$ref": "schemas/audit-page.schema.json"
      },
...
      "Problem": {
        "$ref": "schemas/problem.schema.json"
      }
    }
  }
}